requirements.txt
asgiref==3.2.3
attrs==19.3.0
Django==3.0.4
django-cors-headers==3.2.1
django-pytest==0.2.0
djangorestframework==3.11.0
factory-boy==2.12.0
Faker==4.0.1
importlib-metadata==1.5.0
inflection==0.3.1
more-itertools==8.2.0
packaging==20.3
pluggy==0.13.1
py==1.8.1
pyparsing==2.4.6
pytest==5.3.5
pytest-django==3.8.0
pytest-factoryboy==2.0.3
python-dateutil==2.8.1
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
text-unidecode==1.3
wcwidth==0.1.8
zipp==3.1.0
